Tarea

+++ PubliUcoBusiness/src/co/edu/uco/publiuco/business/business/impl/EstadoTipoRelacionInstitucionBusinessImpl.java
@@ -7,6 +7,8 @@ import co.edu.uco.publiuco.business.assembler.concrete.EstadoTipoRelacionInstitu
 import co.edu.uco.publiuco.business.business.EstadoTipoRelacionInstitucionBusiness;
 import co.edu.uco.publiuco.business.domain.EstadoTipoRelacionInstitucionDomain;
 import co.edu.uco.publiuco.crosscutting.exception.PubliUcoBusinessException;
+import co.edu.uco.publiuco.crosscutting.utils.Messages.EstadoTipoRelacionInstitucionBusinessImplMessage;
+import co.edu.uco.publiuco.crosscutting.utils.UtilText;
 import co.edu.uco.publiuco.crosscutting.utils.UtilUUID;
 import co.edu.uco.publiuco.data.dao.factory.DAOFactory;
 import co.edu.uco.publiuco.entities.EstadoTipoRelacionInstitucionEntity;
@@ -34,10 +36,10 @@ public final class EstadoTipoRelacionInstitucionBusinessImpl implements EstadoTi
 		resultEntities = daoFactory.getEstadoTipoRelacionInstitucionDAO().read(entityTmp);
 		
 		if(!resultEntities.isEmpty()) {
-			throw PubliUcoBusinessException.create("El estado de tipo relacion institucion que intenta crear ya existe, porfavor verific¡que los datos y de ser necesario proceda a actualizar");
+			throw PubliUcoBusinessException.create(EstadoTipoRelacionInstitucionBusinessImplMessage.CREATE_EXCEPTION_PROBLEM_BUSINESS_IMPLEMENTATION_STRING);
 		}
 		
-		final var domainToCreate = new EstadoTipoRelacionInstitucionDomain(idenficador,domain.getNombre(),domain.getDescripcion());
+		//final var domainToCreate = new EstadoTipoRelacionInstitucionDomain(idenficador,domain.getNombre(),domain.getDescripcion());
 		
 		
 		final EstadoTipoRelacionInstitucionEntity entity = EstadoTipoRelacionInstitucionAssembler.getInstance()
@@ -59,6 +61,22 @@ public final class EstadoTipoRelacionInstitucionBusinessImpl implements EstadoTi
 
 	@Override
 	public final void modify(EstadoTipoRelacionInstitucionDomain domain) {
+		String nombre;
+		
+		EstadoTipoRelacionInstitucionEntity entityTmp;
+		List<EstadoTipoRelacionInstitucionEntity> resultEntities;
+		do {
+		nombre = UtilText.getDefaultValue();
+			entityTmp =  EstadoTipoRelacionInstitucionEntity.createWithNombrer(nombre);
+			resultEntities = daoFactory.getEstadoTipoRelacionInstitucionDAO().read(entityTmp);
+		} while (!resultEntities.isEmpty());
+		entityTmp= EstadoTipoRelacionInstitucionEntity.createWithDescripcion(domain.getNombre());
+		resultEntities = daoFactory.getEstadoTipoRelacionInstitucionDAO().read(entityTmp);
+		
+		if(!resultEntities.isEmpty()) {
+			throw PubliUcoBusinessException.create(EstadoTipoRelacionInstitucionBusinessImplMessage.UPDATE_EXCEPTION_PROBLEM_BUSINESS_IMPLEMENTATION_STRING);
+		}
+		
 		final EstadoTipoRelacionInstitucionEntity entity = EstadoTipoRelacionInstitucionAssembler.getInstance()
 				.toEntityFromDomain(domain);
 		daoFactory.getEstadoTipoRelacionInstitucionDAO().update(entity);
@@ -67,6 +85,20 @@ public final class EstadoTipoRelacionInstitucionBusinessImpl implements EstadoTi
 
 	@Override
 	public final void drop(EstadoTipoRelacionInstitucionDomain domain) {
+		UUID idenficador;
+		EstadoTipoRelacionInstitucionEntity entityTmp;
+		List<EstadoTipoRelacionInstitucionEntity> resultEntities;
+		do {
+			idenficador = UtilUUID.genereteNewUUID();
+			entityTmp =  EstadoTipoRelacionInstitucionEntity.createWithIdentificador(idenficador);
+			resultEntities = daoFactory.getEstadoTipoRelacionInstitucionDAO().read(entityTmp);
+		} while (!resultEntities.isEmpty());
+	
+		
+		if(!resultEntities.isEmpty()) {
+			throw PubliUcoBusinessException.create(EstadoTipoRelacionInstitucionBusinessImplMessage.DROP_EXCEPTION_PROBLEM_BUSINESS_IMPLEMENTATION_STRING);
+		}
+		
 		final EstadoTipoRelacionInstitucionEntity entity = EstadoTipoRelacionInstitucionAssembler.getInstance()
 				.toEntityFromDomain(domain);
 		daoFactory.getEstadoTipoRelacionInstitucionDAO().delete(entity);
